requirements.txt
ipykernel
pymysql
sqlalchemy
python-dotenv
cryptography
ipython-sql
pyspark
findspark
mysql-connector-python
pytest